# api/docs/swagger.yaml
definitions:
...
  models.CreateProfession:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  models.Profession:
    properties:
      id:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  models.ResponseModel:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
  models.Status:
    properties:
      status:
        type: string
    type: object
...
  position_service.CreateProfessionRequest:
    properties:
      name:
        type: string
    type: object
...
  position_service.GetAllProfessionResponse:
    properties:
      count:
        type: integer
      professions:
        items:
          $ref: '#/definitions/position_service.Profession'
        type: array
    type: object
...
  position_service.Profession:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/profession:
    get:
      consumes:
      - application/json
      description: get profession
      operationId: get-profession
      parameters:
      - description: search
        in: query
        name: search
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/position_service.GetAllProfessionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get profession all
      tags:
      - profession
    post:
      consumes:
      - application/json
      description: create profession
      operationId: create-professions
      parameters:
      - description: profession
        in: body
        name: profession
        required: true
        schema:
          $ref: '#/definitions/position_service.CreateProfessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/position_service.Profession'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: create profession
      tags:
      - profession
  /v1/profession/{profession_id}:
    delete:
      consumes:
      - application/json
      description: Delete Profession
      operationId: delete-profession
      parameters:
      - description: profession_id
        in: path
        name: profession_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.Status'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: delete profession
      tags:
      - profession
    get:
      consumes:
      - application/json
      description: get profession
      operationId: get-profession-byID
      parameters:
      - description: profession_id
        in: path
        name: profession_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.Profession'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get profession by ID
      tags:
      - profession
    put:
      consumes:
      - application/json
      description: Update Profession by ID
      operationId: update_profession
      parameters:
      - description: profession_id
        in: path
        name: profession_id
        required: true
        type: string
      - description: profession
        in: body
        name: profession
        required: true
        schema:
          $ref: '#/definitions/models.CreateProfession'
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.Status'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Update Profession
      tags:
      - profession
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"